# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Blacket-style game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [player]
      summary: Get current player profile
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    patch:
      operationId: updateMe
      tags: [player]
      summary: Update player settings (avatar blook, banner color)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "400":
          description: Invalid update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /me/claim:
    post:
      operationId: claimTokens
      tags: [player]
      summary: Claim the timed token reward
      responses:
        "200":
          description: Claim result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResult"
        "400":
          description: Claim not ready yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResult"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change your password (requires the current password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "400":
          description: Invalid new password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "401":
          description: Wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /me/blooks/sell:
    post:
      operationId: sellBlooks
      tags: [player]
      summary: Sell owned blooks for tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellRequest"
      responses:
        "200":
          description: Sale result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellResult"
        "400":
          description: Not enough blooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /storage/uploads/request-url:
    post:
      tags: [clans]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [clans]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Player:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        tokens: { type: integer }
        experience: { type: integer }
        level: { type: integer }
        avatarBlook: { type: ["string", "null"] }
        avatarImage: { type: ["string", "null"] }
        bannerColor: { type: ["string", "null"] }
        packsOpened: { type: integer }
        uniqueBlooks: { type: integer }
        totalBlooks: { type: integer }
        nextClaimAt: { type: ["string", "null"] }
        nameEffect: { type: ["string", "null"] }
        chatColor: { type: ["string", "null"], description: "Custom chat text color (#rrggbb) or null for default" }
        hasBundle: { type: boolean, description: "Owns the Starter Bundle (unlocks custom chat/name colors)" }
        clanBoosts: { type: integer, description: "Unused clan level boosts (+10 levels each)" }
        rainbowPerks: { type: integer, description: "Unattached rainbow clan name perks" }
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        createdAt: { type: string }
      required:
        [id, username, tokens, experience, level, avatarBlook, avatarImage, bannerColor, packsOpened, uniqueBlooks, totalBlooks, nextClaimAt, clanBoosts, rainbowPerks, badges, createdAt]
    PlayerUpdate:
      type: object
      properties:
        avatarBlook: { type: string }
        bannerColor: { type: string }
        username: { type: string, minLength: 3, maxLength: 20 }
        currentPassword: { type: string, minLength: 1, maxLength: 100, description: "Required when changing username: the account's current password" }
        chatColor: { type: ["string", "null"], description: "Custom chat text color (#rrggbb); null resets to default (Starter Bundle perk)" }
        nameColor: { type: string, description: "Custom name color (#rrggbb) or 'golden' to switch back (Starter Bundle perk)" }
    ClaimResult:
      type: object
      properties:
        claimed: { type: boolean }
        tokensAwarded: { type: integer }
        tokens: { type: integer }
        nextClaimAt: { type: string }
      required: [claimed, tokensAwarded, tokens, nextClaimAt]
...
    SellRequest:
      type: object
      properties:
        name: { type: string }
        quantity: { type: integer, minimum: 1 }
      required: [name, quantity]
    SellResult:
      type: object
      properties:
        tokensEarned: { type: integer }
        tokens: { type: integer }
        remainingQuantity: { type: integer }
      required: [tokensEarned, tokens, remainingQuantity]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ChangePasswordInput:
      type: object
      properties:
        oldPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8 }
      required: [oldPassword, newPassword]
...
    Badge:
      type: object
      properties:
        name: { type: string }
        image: { type: string }
        description: { type: string }
      required: [name, image, description]
...